Implemented workout list

+++ src/main/java/servlet/swe432.java
@@ -265,7 +265,7 @@ public class swe432 extends HttpServlet {
 		 String gym = request.getParameter(Data.GYM.name());
 		 String experience = request.getParameter(Data.EXPERIENCE.name());
 		 String[] workout = request.getParameterValues(Data.WORKOUT.name());
-		 Integer time = session.getLastAccessedTime().intValue();
+		 Integer time = session.getLastAccessedTime();
 		 String result = "";
 		 if(workout != null) {
 			 if(workout.length == 1)
